requirements.txt
Django==4.0.3
django-extensions==3.1.5
djangorestframework==3.13.1
requests==2.27.1
